// src/app/components/model/cliente.ts
// export interface Cliente{
//     id: number,
//     nome: string,
//     sobreNome: string,
//     telefone: number,
//     cep: number,
//     observacoes: string,
//     localidade: string,
//     email: string
    
// }

export class Cliente{
    id: number;
    nome: string;
    sobreNome: string;
    telefone: number;
    cep: number;
    observacoes: string;
    localidade: string;
    email: string
    
}